Refactor usePanoptesUserGroup with swr
