+++ packages/lib-user/src/hooks/usePanoptesAuth.js
