+++ packages/lib-user/src/hooks/usePanoptesUser.js
@@ -3,6 +3,8 @@ import { useEffect, useState } from 'react'
 
 import { getBearerToken } from '@utils/index.js'
 
+// TODO: refactor with SWR
+
 async function fetchPanoptesUser(authClient) {
   try {
     const authorization = await getBearerToken(authClient)
